add branches list cmd
